python/requirements.txt
# FastAPI 和 Web 服务器
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2

# HTTP 请求
requests==2.32.3

# 机器学习和数据处理
numpy==1.26.4
scikit-learn==1.5.2
umap-learn==0.5.6

# 日志
python-json-logger==2.0.7
